fix build sockets with ipv6 disabled
