+++ aether/transport/system_sockets/sockets/win_tcp_socket.cpp
@@ -24,70 +24,12 @@
 #  include <mswsock.h>
 
 #  include "aether/misc/defer.h"
+#  include "aether/transport/system_sockets/sockets/win_sock_addr.h"
 
 #  include "aether/tele/tele.h"
 
 namespace ae {
-
 namespace win_socket_internal {
-struct SockAddr {
-  sockaddr* addr() { return reinterpret_cast<sockaddr*>(&data); }
-
-  union {
-#  if AE_SUPPORT_IPV4 == 1
-    struct sockaddr_in ipv4;
-#  endif
-#  if AE_SUPPORT_IPV6 == 1
-    struct sockaddr_in6 ipv6;
-#  endif
-  } data;
-
-  std::size_t size;
-};
-
-SockAddr GetSockAddr(IpAddressPort const& ip_address_port) {
-  switch (ip_address_port.ip.version) {
-    case IpAddress::Version::kIpV4:
-#  if AE_SUPPORT_IPV4 == 1
-    {
-      SockAddr sock_addr{};
-      sock_addr.size = sizeof(sock_addr.data.ipv4);
-      auto& addr = sock_addr.data.ipv4;
-
-      std::memcpy(&addr.sin_addr.s_addr, ip_address_port.ip.value.ipv4_value,
-                  4);
-      addr.sin_port = ae::SwapToInet(ip_address_port.port);
-      addr.sin_family = AF_INET;
-
-      return sock_addr;
-    }
-#  else
-    {
-      assert(false);
-      break;
-    }
-#  endif
-    case IpAddress::Version::kIpV6: {
-#  if AE_SUPPORT_IPV6 == 1
-      SockAddr sock_addr;
-      sock_addr.size = sizeof(sock_addr.data.ipv6);
-      auto& addr = sock_addr.data.ipv6;
-      std::memcpy(&addr.sin6_addr, ip_address_port.ip.value.ipv6_value, 16);
-      addr.sin6_port = ae::SwapToInet(ip_address_port.port);
-      addr.sin6_family = AF_INET6;
-
-      return sock_addr;
-    }
-#  else
-      {
-        assert(false);
-        break;
-      }
-#  endif
-  }
-  return {};
-}
-
 LPFN_CONNECTEX GetConnectEx(SOCKET socket) {
   LPFN_CONNECTEX func_ptr = nullptr;
   GUID guid = WSAID_CONNECTEX;
